api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
Pillow>=11.0.0
numpy>=1.26.0
tensorflow-cpu==2.17.0
scikit-learn==1.5.2
streamlit>=1.35.0
requests>=2.32.3
matplotlib>=3.9.0
streamlit-drawable-canvas==0.9.3
